ghvmctl: Try prevent gnome-initial-setup to start
Configure the machine so that initial setup won't start.

Sadly it's not possible to do this by configuring the image earlier, as
trying to mount files while configuring may lead to permission errors.

So just do it as early as possible, keeping a fallback to still stop the
app if this is required