ghvmctl-runner: Duplicate env variables from current user session
Instead of guessing env variables, we should inherit them from current
user session, this can be done getting them from systemd (or checking
the environ of an already known process for the uer), so that we can safely
run in a very similar environment.

This could also be done using systemd-run, although it may have other
implications too.