ghvmctl-runner: Use systemd-run to exec in user env
We can just run a process inside the session with all the right
environment variables, leaving this job to systemd without having to do
extra work.

This seems to do the right job after some testing, so let's just
simplify everything.

Keeping the old commit (instead of squashing), just in case we need to
revert to that one day.